Added Internal DataModel Class for specify (For Forms and Schema Localization)
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2768 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/ui/db/TextFieldFromPickListTable.java
@@ -100,7 +100,7 @@ public class TextFieldFromPickListTable extends JTextField implements GetSetValu
                     if (isFormObjIFace && valObj instanceof FormDataObjIFace)
                     {
                         //System.out.println(((FormDataObjIFace)value).getId().longValue()+"  "+(((FormDataObjIFace)valObj).getId().longValue()));
-                        if (((FormDataObjIFace)value).getId().longValue() == (((FormDataObjIFace)valObj).getId().longValue()))
+                        if (((FormDataObjIFace)value).getId().intValue() == (((FormDataObjIFace)valObj).getId().intValue()))
                         {
                             data = pli.getTitle();
                             break;
